# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/cse/{cluster}/{source}/{event}": {
            "put": {
                "description": "log a source and event as part of a cluster",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "log a source and event as part of a cluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "a UUID representing this cluster",
                        "name": "cluster",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "an event category",
                        "name": "event",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "a source category",
                        "name": "source",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API key, 32 bytes or more, issued",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/itslog.Success"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/itslog.Error"
                        }
                    }
                }
            }
        },
        "/csev/{cluster}/{source}/{event}/{value}": {
            "put": {
                "description": "log a source and event with a unique value as part of a cluster",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "log a source and event with a unique value as part of a cluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "a UUID representing this cluster",
                        "name": "cluster",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "an event category",
                        "name": "event",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "a source category",
                        "name": "source",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "a unique value associated with this event",
                        "name": "value",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API key, 32 bytes or more, issued",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/itslog.Success"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/itslog.Error"
                        }
                    }
                }
            }
        },
        "/se/{source}/{event}": {
            "put": {
                "description": "logs an event with a source",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "log a source/event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "an event category",
                        "name": "event",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "a source category",
                        "name": "source",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API key, 32 bytes or more, issued",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/itslog.Success"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/itslog.Error"
                        }
                    }
                }
            }
        },
        "/sev/{source}/{event}/{value}": {
            "put": {
                "description": "log a source and event with a unique value",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "log a source and event with a unique value",
                "parameters": [
                    {
                        "type": "string",
                        "description": "an event category",
                        "name": "event",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "a source category",
                        "name": "source",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "a unique value associated with this event",
                        "name": "value",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API key, 32 bytes or more, issued",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/itslog.Success"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/itslog.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "itslog.Error": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "error": {
                    "type": "string"
                },
                "error_type": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "itslog.Success": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        }
    }
}